created a new modal generic component and splited searchBookModal
