+++ src/components/common/modal/Modal.css
@@ -31,23 +31,6 @@ div[data-testid='modal_header'] {
   border-bottom: 1px solid #ddd;
 }
 
-div[data-testid='modal_body'] {
-  display: flex;
-  flex-direction: column;
-  padding: 1rem;
-  overflow-y: auto;
-  gap: 1rem;
-}
-
-div[data-testid^='modal_book_item'] {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  border-bottom: 1px solid #ddd;
-  gap: 1rem;
-  padding: 1rem;
-}
-
 @media screen and (max-width: 650px) {
   div[data-testid='modal_component'] {
     width: 95%;
