+++ src/components/searchBookSection/searchBookModal/SearchBookModal.css
@@ -1,0 +1,16 @@
+div[data-testid='modal_body'] {
+  display: flex;
+  flex-direction: column;
+  padding: 1rem;
+  overflow-y: auto;
+  gap: 1rem;
+}
+
+div[data-testid^='modal_book_item'] {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  border-bottom: 1px solid #ddd;
+  gap: 1rem;
+  padding: 1rem;
+}
